requirements.txt
numpy
pandas
scikit-learn
pyarrow
joblib
tqdm
matplotlib
seaborn
tensorflow-macos
tensorflow-metal

